resume-editor-backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pdfplumber
python-docx
docx2txt
